server/storage/cherrypy/src/mtt_server_requirements.txt
CherryPy==3.8.0
configobj==5.0.6
nose==1.3.7
pycrypto==2.6.1
requests==2.7.0
six==1.9.0
wheel==0.24.0
psycopg2==2.6.1
